backend/requirements.txt
fastapi==0.115.12
PyJWT==2.10.1
uvicorn[standard]==0.34.0
numpy>=2.2.0
pandas>=2.2.0
torch
scapy
psycopg[binary]>=3.2.0
elasticsearch>=8.15.0